pyproject.toml
[project]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniil Korobko",email = "“korobko_01@inbox.ru”"}
]
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "django (>=2.2,<6.0)",
    "psycopg (>=3.3.2,<4.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "ruff (>=0.14.9,<0.15.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "djangorestframework-types (>=0.9.0,<0.10.0)",
    "django-stubs (>=5.2.8,<6.0.0)",
    "mypy-extensions (>=1.1.0,<2.0.0)",
    "pydantic[dotenv] (>=2.12.5,<3.0.0)",
    "aiogram (>=3.23.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "django-filter (>=25.2,<26.0)",
    "celery (>=5.6.2,<6.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "types-pytz (>=2025.2.0.20251108,<2026.0.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
